+++ tango/__main__.py
@@ -90,6 +90,8 @@ def run(
 ):
     """
     Run a tango experiment.
+
+    EXPERIMENT is the path to experiment's JSON/JSonnet/YAML configuration file.
     """
     _run(
         experiment,